@@ -1,10 +1,13 @@
-// File generated by FlutterFire CLI.
+// lib/firebase_options.dart
+// File generated by FlutterFire CLI (actualizado manualmente para bucket correcto).
 // ignore_for_file: type=lint
+
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
 
 class DefaultFirebaseOptions {
+  /// Retorna la configuración de Firebase según la plataforma actual.
   static FirebaseOptions get currentPlatform {
     if (kIsWeb) {
       return web;
@@ -36,29 +39,45 @@ class DefaultFirebaseOptions {
     }
   }
 
+  /// 🔹 NUEVO: Getter práctico para obtener el bucket correcto donde sea necesario
+  static String get storageBucketName =>
+      currentPlatform.storageBucket ?? 'fauna2admin.firebasestorage.app';
+
+  // =====================================================
+  // Configuración para Web
+  // =====================================================
   static const FirebaseOptions web = FirebaseOptions(
     apiKey: 'AIzaSyCFPOy6zdvIFZxLx1bQEavRC-pgxFztZyg',
     appId: '1:190654239142:web:e5dc22baed4a895944bf67',
     messagingSenderId: '190654239142',
     projectId: 'fauna2admin',
     authDomain: 'fauna2admin.firebaseapp.com',
-    storageBucket: 'fauna2admin.appspot.com', // ✅ corregido
+    storageBucket:
+    'fauna2admin.firebasestorage.app', // ✅ corregido: .firebasestorage.app
   );
 
+  // =====================================================
+  // Configuración para Android
+  // =====================================================
   static const FirebaseOptions android = FirebaseOptions(
     apiKey: 'AIzaSyBxvxeWyp9TQjKqpfpk8Nz-dxjhnhKFxO8',
     appId: '1:190654239142:android:c99ec8cf9f8d1bc144bf67',
     messagingSenderId: '190654239142',
     projectId: 'fauna2admin',
-    storageBucket: 'fauna2admin.appspot.com', // ✅ corregido
+    storageBucket:
+    'fauna2admin.firebasestorage.app', // ✅ corregido: .firebasestorage.app
   );
 
+  // =====================================================
+  // Configuración para iOS
+  // =====================================================
   static const FirebaseOptions ios = FirebaseOptions(
     apiKey: 'AIzaSyAsSOvvAawBVVadThaBHjwnxeNVyfHz09w',
     appId: '1:190654239142:ios:e6a39200f362cb5444bf67',
     messagingSenderId: '190654239142',
     projectId: 'fauna2admin',
-    storageBucket: 'fauna2admin.appspot.com', // ✅ corregido
+    storageBucket:
+    'fauna2admin.firebasestorage.app', // ✅ corregido: .firebasestorage.app
     iosBundleId: 'com.example.faunadmin2',
   );
 }
